+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When key snippets are emitted, snippet files are auto added to related files.
 - Add base task to research subtask prompt.
 - Adjust research prompt to make sure related files are related to the base task, not just the research subtask.
+- Track tasks by ID and allow them to be deleted.
+- Make one_shot_completed tool available to research agent.
 
 ## [0.6.0] - 2024-12-17